Accordion Group update
